Moving ClusterConfig from job.manager.core to stream.processor.common
