+++ components/org.wso2.carbon.sp.jobmanager.core/src/test/java/org/wso2/carbon/sp/jobmanager/core/BeanTestCase.java
@@ -20,10 +20,10 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.wso2.carbon.sp.jobmanager.core.appcreator.DeployableSiddhiQueryGroup;
 import org.wso2.carbon.sp.jobmanager.core.appcreator.DistributedSiddhiQuery;
-import org.wso2.carbon.sp.jobmanager.core.bean.ClusterConfig;
+import org.wso2.carbon.stream.processor.common.utils.config.ClusterConfig;
 import org.wso2.carbon.sp.jobmanager.core.bean.DeploymentConfig;
 import org.wso2.carbon.sp.jobmanager.core.bean.InterfaceConfig;
-import org.wso2.carbon.sp.jobmanager.core.bean.StrategyConfig;
+import org.wso2.carbon.stream.processor.common.utils.config.StrategyConfig;
 import org.wso2.carbon.sp.jobmanager.core.model.ManagerNode;
 import org.wso2.carbon.sp.jobmanager.core.model.ManagerNodeConfig;
 import org.wso2.carbon.sp.jobmanager.core.model.ResourceNode;
